Remove access.Signup
Move the function to the server package, since it does not perform any authorization
And rename it so that it's not confused with the existing Signup endpoint.
